requirements.txt
# Hotel Search Agent Requirements

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# AI & LangChain
openai==1.54.3
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.17

# Perplexity SDK
perplexityai==0.22.0

# Google Places API
googlemaps==4.10.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiohttp==3.11.7
tenacity==9.0.0

# Logging & Monitoring
rich==13.9.4
loguru==0.7.2

# Date handling
python-dateutil==2.9.0
